requirements.txt
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
fsspec==2024.12.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onvif_zeep==0.2.12
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
zeep==4.3.1
